refactor: extract commands
